Refactor patch filtering on the path towards making it work properly
