+++ src/wix/WixToolset.Core.WindowsInstaller/Bind/CreatePatchSubStoragesCommand.cs
@@ -17,7 +17,7 @@ namespace WixToolset.Core.WindowsInstaller.Bind
     /// <summary>
     /// Include transforms in a patch.
     /// </summary>
-    internal class AttachPatchTransformsCommand
+    internal class CreatePatchSubStoragesCommand
     {
         private static readonly string[] PatchUninstallBreakingTables = new[]
         {
@@ -55,7 +55,7 @@ namespace WixToolset.Core.WindowsInstaller.Bind
 
         private readonly TableDefinitionCollection tableDefinitions;
 
-        public AttachPatchTransformsCommand(IMessaging messaging, IBackendHelper backendHelper, Intermediate intermediate, IEnumerable<PatchTransform> transforms)
+        public CreatePatchSubStoragesCommand(IMessaging messaging, IBackendHelper backendHelper, Intermediate intermediate, IEnumerable<PatchTransform> transforms)
         {
             this.tableDefinitions = new TableDefinitionCollection(WindowsInstallerTableDefinitions.All);
             this.Messaging = messaging;
@@ -126,22 +126,11 @@ namespace WixToolset.Core.WindowsInstaller.Bind
 
             foreach (var mainTransform in this.Transforms)
             {
-                var baselineSymbol = baselineSymbolsById[mainTransform.Baseline];
-
-                var patchRefSymbols = symbols.OfType<WixPatchRefSymbol>().ToList();
-                if (patchRefSymbols.Count > 0)
-                {
-                    if (!this.ReduceTransform(mainTransform.Transform, patchRefSymbols))
-                    {
-                        // transform has none of the content authored into this patch
-                        continue;
-                    }
-                }
-
                 // Validate the transform doesn't break any patch specific rules.
                 this.Validate(mainTransform);
 
                 // ensure consistent File.Sequence within each Media
+                var baselineSymbol = baselineSymbolsById[mainTransform.Baseline];
                 var mediaSymbol = patchMediaByDiskId[baselineSymbol.DiskId];
 
                 // Ensure that files are sequenced after the last file in any transform.
@@ -383,331 +372,9 @@ namespace WixToolset.Core.WindowsInstaller.Bind
             return success;
         }
 
-        /// <summary>
-        /// Reduce the transform according to the patch references.
-        /// </summary>
-        /// <param name="transform">transform generated by torch.</param>
-        /// <param name="patchRefSymbols">Table contains patch family filter.</param>
-        /// <returns>true if the transform is not empty</returns>
-        private bool ReduceTransform(WindowsInstallerData transform, IEnumerable<WixPatchRefSymbol> patchRefSymbols)
-        {
-            // identify sections to keep
-            var oldSections = new Dictionary<string, Row>();
-            var newSections = new Dictionary<string, Row>();
-            var tableKeyRows = new Dictionary<string, Dictionary<string, Row>>();
-            var sequenceList = new List<Table>();
-            var componentFeatureAddsIndex = new Dictionary<string, List<string>>();
-            var customActionTable = new Dictionary<string, Row>();
-            var directoryTableAdds = new Dictionary<string, Row>();
-            var featureTableAdds = new Dictionary<string, Row>();
-            var keptComponents = new Dictionary<string, Row>();
-            var keptDirectories = new Dictionary<string, Row>();
-            var keptFeatures = new Dictionary<string, Row>();
-            var keptLockPermissions = new HashSet<string>();
-            var keptMsiLockPermissionExs = new HashSet<string>();
-
-            var componentCreateFolderIndex = new Dictionary<string, List<string>>();
-            var directoryLockPermissionsIndex = new Dictionary<string, List<Row>>();
-            var directoryMsiLockPermissionsExIndex = new Dictionary<string, List<Row>>();
-
-            foreach (var patchRefSymbol in patchRefSymbols)
-            {
-                var tableName = patchRefSymbol.Table;
-                var key = patchRefSymbol.PrimaryKeys;
-
-                // Short circuit filtering if all changes should be included.
-                if ("*" == tableName && "*" == key)
-                {
-                    RemoveProductCodeFromTransform(transform);
-                    return true;
-                }
-
-                if (!transform.Tables.TryGetTable(tableName, out var table))
-                {
-                    // Table not found.
-                    continue;
-                }
-
-                // Index the table.
-                if (!tableKeyRows.TryGetValue(tableName, out var keyRows))
-                {
-                    keyRows = new Dictionary<string, Row>();
-                    tableKeyRows.Add(tableName, keyRows);
-
-                    foreach (var newRow in table.Rows)
-                    {
-                        var primaryKey = newRow.GetPrimaryKey();
-                        keyRows.Add(primaryKey, newRow);
-                    }
-                }
-
-                if (!keyRows.TryGetValue(key, out var row))
-                {
-                    // Row not found.
-                    continue;
-                }
-
-                // Differ.sectionDelimiter
-                var sections = row.SectionId.Split('/');
-                oldSections[sections[0]] = row;
-                newSections[sections[1]] = row;
-            }
-
-            // throw away sections not referenced
-            var keptRows = 0;
-            Table directoryTable = null;
-            Table featureTable = null;
-            Table lockPermissionsTable = null;
-            Table msiLockPermissionsTable = null;
-
-            foreach (var table in transform.Tables)
-            {
-                if ("_SummaryInformation" == table.Name)
-                {
-                    continue;
-                }
-
-                if (table.Name == "AdminExecuteSequence"
-                    || table.Name == "AdminUISequence"
-                    || table.Name == "AdvtExecuteSequence"
-                    || table.Name == "InstallUISequence"
-                    || table.Name == "InstallExecuteSequence")
-                {
-                    sequenceList.Add(table);
-                    continue;
-                }
-
-                for (var i = 0; i < table.Rows.Count; i++)
-                {
-                    var row = table.Rows[i];
-
-                    if (table.Name == "CreateFolder")
-                    {
-                        var createFolderComponentId = row.FieldAsString(1);
-
-                        if (!componentCreateFolderIndex.TryGetValue(createFolderComponentId, out var directoryList))
-                        {
-                            directoryList = new List<string>();
-                            componentCreateFolderIndex.Add(createFolderComponentId, directoryList);
-                        }
-
-                        directoryList.Add(row.FieldAsString(0));
-                    }
-
-                    if (table.Name == "CustomAction")
-                    {
-                        customActionTable.Add(row.FieldAsString(0), row);
-                    }
-
-                    if (table.Name == "Directory")
-                    {
-                        directoryTable = table;
-                        if (RowOperation.Add == row.Operation)
-                        {
-                            directoryTableAdds.Add(row.FieldAsString(0), row);
-                        }
-                    }
-
-                    if (table.Name == "Feature")
-                    {
-                        featureTable = table;
-                        if (RowOperation.Add == row.Operation)
-                        {
-                            featureTableAdds.Add(row.FieldAsString(0), row);
-                        }
-                    }
-
-                    if (table.Name == "FeatureComponents")
-                    {
-                        if (RowOperation.Add == row.Operation)
-                        {
-                            var featureId = row.FieldAsString(0);
-                            var componentId = row.FieldAsString(1);
-
-                            if (!componentFeatureAddsIndex.TryGetValue(componentId, out var featureList))
-                            {
-                                featureList = new List<string>();
-                                componentFeatureAddsIndex.Add(componentId, featureList);
-                            }
-
-                            featureList.Add(featureId);
-                        }
-                    }
-
-                    if (table.Name == "LockPermissions")
-                    {
-                        lockPermissionsTable = table;
-                        if ("CreateFolder" == row.FieldAsString(1))
-                        {
-                            var directoryId = row.FieldAsString(0);
-
-                            if (!directoryLockPermissionsIndex.TryGetValue(directoryId, out var rowList))
-                            {
-                                rowList = new List<Row>();
-                                directoryLockPermissionsIndex.Add(directoryId, rowList);
-                            }
-
-                            rowList.Add(row);
-                        }
-                    }
-
-                    if (table.Name == "MsiLockPermissionsEx")
-                    {
-                        msiLockPermissionsTable = table;
-                        if ("CreateFolder" == row.FieldAsString(1))
-                        {
-                            var directoryId = row.FieldAsString(0);
-
-                            if (!directoryMsiLockPermissionsExIndex.TryGetValue(directoryId, out var rowList))
-                            {
-                                rowList = new List<Row>();
-                                directoryMsiLockPermissionsExIndex.Add(directoryId, rowList);
-                            }
-
-                            rowList.Add(row);
-                        }
-                    }
-
-                    if (null == row.SectionId)
-                    {
-                        table.Rows.RemoveAt(i);
-                        i--;
-                    }
-                    else
-                    {
-                        var sections = row.SectionId.Split('/');
-                        // ignore the row without section id.
-                        if (0 == sections[0].Length && 0 == sections[1].Length)
-                        {
-                            table.Rows.RemoveAt(i);
-                            i--;
-                        }
-                        else if (IsInPatchFamily(sections[0], sections[1], oldSections, newSections))
-                        {
-                            if ("Component" == table.Name)
-                            {
-                                keptComponents.Add(row.FieldAsString(0), row);
-                            }
-
-                            if ("Directory" == table.Name)
-                            {
-                                keptDirectories.Add(row.FieldAsString(0), row);
-                            }
-
-                            if ("Feature" == table.Name)
-                            {
-                                keptFeatures.Add(row.FieldAsString(0), row);
-                            }
-
-                            keptRows++;
-                        }
-                        else
-                        {
-                            table.Rows.RemoveAt(i);
-                            i--;
-                        }
-                    }
-                }
-            }
-
-            keptRows += ReduceTransformSequenceTable(sequenceList, oldSections, newSections, customActionTable);
-
-            if (null != directoryTable)
-            {
-                foreach (var componentRow in keptComponents.Values)
-                {
-                    var componentId = componentRow.FieldAsString(0);
-
-                    if (RowOperation.Add == componentRow.Operation)
-                    {
-                        // Make sure each added component has its required directory and feature heirarchy.
-                        var directoryId = componentRow.FieldAsString(2);
-                        while (null != directoryId && directoryTableAdds.TryGetValue(directoryId, out var directoryRow))
-                        {
-                            if (!keptDirectories.ContainsKey(directoryId))
-                            {
-                                directoryTable.Rows.Add(directoryRow);
-                                keptDirectories.Add(directoryId, directoryRow);
-                                keptRows++;
-                            }
-
-                            directoryId = directoryRow.FieldAsString(1);
-                        }
-
-                        if (componentFeatureAddsIndex.TryGetValue(componentId, out var componentFeatureIds))
-                        {
-                            foreach (var featureId in componentFeatureIds)
-                            {
-                                var currentFeatureId = featureId;
-                                while (null != currentFeatureId && featureTableAdds.TryGetValue(currentFeatureId, out var featureRow))
-                                {
-                                    if (!keptFeatures.ContainsKey(currentFeatureId))
-                                    {
-                                        featureTable.Rows.Add(featureRow);
-                                        keptFeatures.Add(currentFeatureId, featureRow);
-                                        keptRows++;
-                                    }
-
-                                    currentFeatureId = featureRow.FieldAsString(1);
-                                }
-                            }
-                        }
-                    }
-
-                    // Hook in changes LockPermissions and MsiLockPermissions for folders for each component that has been kept.
-                    foreach (var keptComponentId in keptComponents.Keys)
-                    {
-                        if (componentCreateFolderIndex.TryGetValue(keptComponentId, out var directoryList))
-                        {
-                            foreach (var directoryId in directoryList)
-                            {
-                                if (directoryLockPermissionsIndex.TryGetValue(directoryId, out var lockPermissionsRowList))
-                                {
-                                    foreach (var lockPermissionsRow in lockPermissionsRowList)
-                                    {
-                                        var key = lockPermissionsRow.GetPrimaryKey('/');
-                                        if (keptLockPermissions.Add(key))
-                                        {
-                                            lockPermissionsTable.Rows.Add(lockPermissionsRow);
-                                            keptRows++;
-                                        }
-                                    }
-                                }
-
-                                if (directoryMsiLockPermissionsExIndex.TryGetValue(directoryId, out var msiLockPermissionsExRowList))
-                                {
-                                    foreach (var msiLockPermissionsExRow in msiLockPermissionsExRowList)
-                                    {
-                                        var key = msiLockPermissionsExRow.GetPrimaryKey('/');
-                                        if (keptMsiLockPermissionExs.Add(key))
-                                        {
-                                            msiLockPermissionsTable.Rows.Add(msiLockPermissionsExRow);
-                                            keptRows++;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            keptRows += ReduceTransformSequenceTable(sequenceList, oldSections, newSections, customActionTable);
-
-            // Delete tables that are empty.
-            var tablesToDelete = transform.Tables.Where(t => t.Rows.Count == 0).Select(t => t.Name).ToList();
-
-            foreach (var tableName in tablesToDelete)
-            {
-                transform.Tables.Remove(tableName);
-            }
-
-            return keptRows > 0;
-        }
-
         private void Validate(PatchTransform patchTransform)
         {
-            var transformPath = patchTransform.Baseline; // TODO: this is used in error messages, how best to set it?
+            var transformPath = patchTransform.Baseline;
             var transform = patchTransform.Transform;
 
             // Changing the ProdocutCode in a patch transform is not recommended.
@@ -905,194 +572,6 @@ namespace WixToolset.Core.WindowsInstaller.Bind
                     }
                 }
             }
-        }
-
-        /// <summary>
-        /// Remove the ProductCode property from the transform.
-        /// </summary>
-        /// <param name="transform">The transform.</param>
-        /// <remarks>
-        /// Changing the ProductCode is not supported in a patch.
-        /// </remarks>
-        private static void RemoveProductCodeFromTransform(WindowsInstallerData transform)
-        {
-            if (transform.Tables.TryGetTable("Property", out var propertyTable))
-            {
-                for (var i = 0; i < propertyTable.Rows.Count; ++i)
-                {
-                    var propertyRow = propertyTable.Rows[i];
-                    var property = propertyRow.FieldAsString(0);
-
-                    if ("ProductCode" == property)
-                    {
-                        propertyTable.Rows.RemoveAt(i);
-                        break;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Check if the section is in a PatchFamily.
-        /// </summary>
-        /// <param name="oldSection">Section id in target wixout</param>
-        /// <param name="newSection">Section id in upgrade wixout</param>
-        /// <param name="oldSections">Dictionary contains section id should be kept in the baseline wixout.</param>
-        /// <param name="newSections">Dictionary contains section id should be kept in the upgrade wixout.</param>
-        /// <returns>true if section in patch family</returns>
-        private static bool IsInPatchFamily(string oldSection, string newSection, Dictionary<string, Row> oldSections, Dictionary<string, Row> newSections)
-        {
-            var result = false;
-
-            if ((String.IsNullOrEmpty(oldSection) && newSections.ContainsKey(newSection)) || (String.IsNullOrEmpty(newSection) && oldSections.ContainsKey(oldSection)))
-            {
-                result = true;
-            }
-            else if (!String.IsNullOrEmpty(oldSection) && !String.IsNullOrEmpty(newSection) && (oldSections.ContainsKey(oldSection) || newSections.ContainsKey(newSection)))
-            {
-                result = true;
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Reduce the transform sequence tables.
-        /// </summary>
-        /// <param name="sequenceList">ArrayList of tables to be reduced</param>
-        /// <param name="oldSections">Hashtable contains section id should be kept in the baseline wixout.</param>
-        /// <param name="newSections">Hashtable contains section id should be kept in the target wixout.</param>
-        /// <param name="customAction">Hashtable contains all the rows in the CustomAction table.</param>
-        /// <returns>Number of rows left</returns>
-        private static int ReduceTransformSequenceTable(List<Table> sequenceList, Dictionary<string, Row> oldSections, Dictionary<string, Row> newSections, Dictionary<string, Row> customAction)
-        {
-            var keptRows = 0;
-
-            foreach (var currentTable in sequenceList)
-            {
-                for (var i = 0; i < currentTable.Rows.Count; i++)
-                {
-                    var row = currentTable.Rows[i];
-                    var actionName = row.Fields[0].Data.ToString();
-                    var sections = row.SectionId.Split('/');
-                    var isSectionIdEmpty = (sections[0].Length == 0 && sections[1].Length == 0);
-
-                    if (row.Operation == RowOperation.None)
-                    {
-                        // Ignore the rows without section id.
-                        if (isSectionIdEmpty)
-                        {
-                            currentTable.Rows.RemoveAt(i);
-                            i--;
-                        }
-                        else if (IsInPatchFamily(sections[0], sections[1], oldSections, newSections))
-                        {
-                            keptRows++;
-                        }
-                        else
-                        {
-                            currentTable.Rows.RemoveAt(i);
-                            i--;
-                        }
-                    }
-                    else if (row.Operation == RowOperation.Modify)
-                    {
-                        var sequenceChanged = row.Fields[2].Modified;
-                        var conditionChanged = row.Fields[1].Modified;
-
-                        if (sequenceChanged && !conditionChanged)
-                        {
-                            keptRows++;
-                        }
-                        else if (!sequenceChanged && conditionChanged)
-                        {
-                            if (isSectionIdEmpty)
-                            {
-                                currentTable.Rows.RemoveAt(i);
-                                i--;
-                            }
-                            else if (IsInPatchFamily(sections[0], sections[1], oldSections, newSections))
-                            {
-                                keptRows++;
-                            }
-                            else
-                            {
-                                currentTable.Rows.RemoveAt(i);
-                                i--;
-                            }
-                        }
-                        else if (sequenceChanged && conditionChanged)
-                        {
-                            if (isSectionIdEmpty)
-                            {
-                                row.Fields[1].Modified = false;
-                                keptRows++;
-                            }
-                            else if (IsInPatchFamily(sections[0], sections[1], oldSections, newSections))
-                            {
-                                keptRows++;
-                            }
-                            else
-                            {
-                                row.Fields[1].Modified = false;
-                                keptRows++;
-                            }
-                        }
-                    }
-                    else if (row.Operation == RowOperation.Delete)
-                    {
-                        if (isSectionIdEmpty)
-                        {
-                            // it is a stardard action which is added by wix, we should keep this action.
-                            row.Operation = RowOperation.None;
-                            keptRows++;
-                        }
-                        else if (IsInPatchFamily(sections[0], sections[1], oldSections, newSections))
-                        {
-                            keptRows++;
-                        }
-                        else
-                        {
-                            if (customAction.ContainsKey(actionName))
-                            {
-                                currentTable.Rows.RemoveAt(i);
-                                i--;
-                            }
-                            else
-                            {
-                                // it is a stardard action, we should keep this action.
-                                row.Operation = RowOperation.None;
-                                keptRows++;
-                            }
-                        }
-                    }
-                    else if (row.Operation == RowOperation.Add)
-                    {
-                        if (isSectionIdEmpty)
-                        {
-                            keptRows++;
-                        }
-                        else if (IsInPatchFamily(sections[0], sections[1], oldSections, newSections))
-                        {
-                            keptRows++;
-                        }
-                        else
-                        {
-                            if (customAction.ContainsKey(actionName))
-                            {
-                                currentTable.Rows.RemoveAt(i);
-                                i--;
-                            }
-                            else
-                            {
-                                keptRows++;
-                            }
-                        }
-                    }
-                }
-            }
-
-            return keptRows;
         }
 
         /// <summary>
